--- atlas_rag/billion/retriever/__init__.py
@@ -1,1 +1,0 @@
-from .lkgr import LargeKGRetriever
