+++ atlas_rag/retriever/hipporag.py
@@ -2,7 +2,7 @@ from tqdm import tqdm
 import networkx as nx
 import numpy as np
 from atlas_rag.retrieval.embedding_model import BaseEmbeddingModel
-from atlas_rag.reader.llm_generator import LLMGenerator
+from atlas_rag.llm_generator.llm_generator import LLMGenerator
 from logging import Logger
 from typing import Optional
 from atlas_rag.retrieval.retriever.base import BasePassageRetriever, InferenceConfig
