+++ atlas_rag/retriever/hipporag2.py
@@ -7,7 +7,7 @@ from typing import Dict, List, Tuple
 import networkx as nx
 import numpy as np
 from atlas_rag.retrieval.embedding_model import BaseEmbeddingModel
-from atlas_rag.reader.llm_generator import LLMGenerator
+from atlas_rag.llm_generator.llm_generator import LLMGenerator
 from logging import Logger
 from dataclasses import dataclass
 from typing import Optional
