+++ atlas_rag/retriever/lkg_retriever/lkgr.py
@@ -7,10 +7,10 @@ import nltk
 from nltk.corpus import stopwords
 nltk.download('stopwords')
 from graphdatascience import GraphDataScience
-from atlas_rag.reader.llm_generator import LLMGenerator
+from atlas_rag.llm_generator.llm_generator import LLMGenerator
 from atlas_rag.retrieval.embedding_model import BaseEmbeddingModel
 import string
-from atlas_rag.billion.retriever.base import BaseLargeKGRetriever
+from atlas_rag.billion.lkg_retriever.base import BaseLargeKGRetriever
