+++ atlas_rag/retriever/tog.py
@@ -1,6 +1,6 @@
 import numpy as np
 from atlas_rag.retrieval.embedding_model import BaseEmbeddingModel
-from atlas_rag.reader.llm_generator import LLMGenerator
+from atlas_rag.llm_generator.llm_generator import LLMGenerator
 from typing import Optional
 from atlas_rag.retrieval.retriever.base import BaseEdgeRetriever, InferenceConfig
