+++ atlas_rag/vectorstore/embedding_model.py
